backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.5
pandas==2.2.3
openpyxl==3.1.2
pdfplumber==0.7.6
PyMuPDF==1.22.5
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
